secured delete

+++ src/app.js
@@ -272,12 +272,14 @@ class App {
 
     function commentDelete(e){
       //event propagation
-      if (e.target.className === 'delete' && e.target.parentNode.dataset['userId'] === User.getCurrentUser().id){
+      if (e.target.className === 'delete' && User.getCurrentUser() !== null && e.target.parentNode.dataset['userId'] == User.getCurrentUser().id){
         //delete comment node
         let commentId = parseInt(e.target.parentNode.dataset['commentId'])
         e.target.parentNode.remove()
         // remove comment
         Adapter.deleteComment(commentId)
+      } else if(e.target.className ==='delete') {
+        alert('You can only delete your own comments')
       }
         e.stopPropagation()
     }
